Getting Started

An overview of the usage of ReadStatTables.jl is provided below. For instructions on installation, see Installation.

Reading a Data File

Suppose we have a Stata .dta file located at data/sample.dta. To read this file into Julia:

julia> using ReadStatTables
julia> tb = readstat("data/sample.dta")5×7 ReadStatTable:
 Row │ mychar    mynum      mydate                dtime         mylabl         ⋯
     │ String  Float64       Date?            DateTime?  Labeled{Int8}  Labele ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │      a      1.1  2018-05-06  2018-05-06T10:10:10           Male         ⋯
   2 │      b      1.2  1880-05-06  1880-05-06T10:10:10         Female         ⋯
   3 │      c  -1000.3  1960-01-01  1960-01-01T00:00:00           Male         ⋯
   4 │      d     -1.4  1583-01-01  1583-01-01T00:00:00         Female         ⋯
   5 │      e   1000.3     missing              missing           Male         ⋯
                                                               2 columns omitted

Here is how we read the above result:[1]

Variable names from the data file are displayed in the first row.

Element type of each variable is displayed below the corresponding variable name.

The values of each variable are displayed column-wise starting from the third row.

Some additional details to be noted:

If a variable contains any missing value, there is a question mark ? in the displayed element type.

By default, all missing values are treated as missing, a special value in Julia.

The date and time values have been translated into Date and DateTime respectively.[2]

Labels instead of the numeric values are displayed for variables with value labels.

Labeled{Int8} is an abbreviation for LabeledValue{Int8}.

Accessing Individual Objects

A vector of all variable names can be obtained as follows:

julia> columnnames(tb)7-element Vector{Symbol}:
 :mychar
 :mynum
 :mydate
 :dtime
 :mylabl
 :myord
 :mytime

To retrieve the array containing data for a specific variable:

julia> tb.mylabl5-element LabeledVector{Union{Missing, Int8}, Vector{Union{Missing, Int8}}, Union{Char, Int32}}:
 1 => Male
 2 => Female
 1 => Male
 2 => Female
 1 => Male

Note

The returned array is exactly the same array holding the data for the table. Therefore, modifying elements in the returned array will also change the data in the table. To avoid such changes, please copy the array first.

Metadata for the data file can be accessed from tb using methods that are compatible with DataAPI.jl.

julia> metadata(tb)ReadStatMeta:
  row count           => 5
  var count           => 7
  modified time       => 2021-04-22T21:36:00
  file format version => 118
  file label          => A test file
  file extension      => .dta
julia> colmetadata(tb)ColMetaIterator{ReadStatColMeta} with 7 entries:
  :mychar => ReadStatColMeta(character, %-1s)
  :mynum  => ReadStatColMeta(numeric, %16.2f)
  :mydate => ReadStatColMeta(date, %td)
  :dtime  => ReadStatColMeta(datetime, %tc)
  :mylabl => ReadStatColMeta(labeled, %16.0f)
  :myord  => ReadStatColMeta(ordinal, %16.0f)
  :mytime => ReadStatColMeta(time, %tcHH:MM:SS)
julia> colmetadata(tb, :myord)ReadStatColMeta:
  label         => ordinal
  format        => %16.0f
  type          => READSTAT_TYPE_INT8
  value label   => myord
  storage width => 1
  display width => 16
  measure       => READSTAT_MEASURE_UNKNOWN
  alignment     => READSTAT_ALIGNMENT_RIGHT

Type Conversions

The interface provided by ReadStatTables.jl allows basic tasks. In case more complicated operations are needed, it is easy to convert the objects into other types.

Converting ReadStatTable

The table returned by readstat is a ReadStatTable. Converting a ReadStatTable to another table type is easy, thanks to the widely supported Tables.jl interface.

For example, to convert a ReadStatTable to a DataFrame from DataFrames.jl:

julia> using DataFrames
julia> df = DataFrame(tb)5×7 DataFrame
 Row │ mychar  mynum    mydate      dtime                mylabl     myord      ⋯
     │ String  Float64  Date?       DateTime?            LabeledV…  LabeledV…  ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │ a           1.1  2018-05-06  2018-05-06T10:10:10  Male       low        ⋯
   2 │ b           1.2  1880-05-06  1880-05-06T10:10:10  Female     medium
   3 │ c       -1000.3  1960-01-01  1960-01-01T00:00:00  Male       high
   4 │ d          -1.4  1583-01-01  1583-01-01T00:00:00  Female     low
   5 │ e        1000.3  missing     missing              Male       missing    ⋯
                                                                1 column omitted

Metadata contained in a ReadStatTable are preserved in the converted DataFrame when working with DataFrames.jl version v1.4.0 or above, which supports the same DataAPI.jl interface for metadata:

julia> metadata(df)Dict{String, Any} with 13 entries:
  "file_ext"             => ".dta"
  "modified_time"        => DateTime("2021-04-22T21:36:00")
  "file_format_version"  => 118
  "file_format_is_64bit" => true
  "table_name"           => ""
  "notes"                => String[]
  "file_encoding"        => ""
  "file_label"           => "A test file"
  "var_count"            => 7
  "row_count"            => 5
  "creation_time"        => DateTime("2021-04-22T21:36:00")
  "endianness"           => READSTAT_ENDIAN_LITTLE
  "compression"          => READSTAT_COMPRESS_NONE
julia> colmetadata(df, :myord)Dict{String, Any} with 8 entries:
  "label"         => "ordinal"
  "format"        => "%16.0f"
  "display_width" => 16
  "measure"       => READSTAT_MEASURE_UNKNOWN
  "alignment"     => READSTAT_ALIGNMENT_RIGHT
  "type"          => READSTAT_TYPE_INT8
  "storage_width" => 0x0000000000000001
  "vallabel"      => :myord

Converting LabeledArray

Variables with value labels are stored in LabeledArrays. To convert a LabeledArray to another array type, we may either obtain an array of LabeledValues or collect the values and labels separately. The data values can be directly retrieved by calling refarray:

julia> refarray(tb.mylabl)5-element Vector{Union{Missing, Int8}}:
 1
 2
 1
 2
 1

Note

The array returned by refarray is exactly the same array underlying the LabeledArray. Therefore, modifying the elements of the array will also mutate the values in the associated LabeledArray.

If only the value labels are needed, we can obtain an iterator of the value labels via valuelabels. For example, to convert a LabeledArray to a CategoricalArray from CategoricalArrays.jl:

julia> using CategoricalArrays
julia> CategoricalArray(valuelabels(tb.mylabl))5-element CategoricalArray{String,1,UInt32}:
 "Male"
 "Female"
 "Male"
 "Female"
 "Male"

It is also possible to only convert the type of the underlying data values:

julia> convertvalue(Int32, tb.mylabl)5-element LabeledVector{Int32, Vector{Int32}, Union{Char, Int32}}:
 1 => Male
 2 => Female
 1 => Male
 2 => Female
 1 => Male

ReadStatTables.convertvalue — Function

convertvalue(T, x::LabeledArray)

Convert the type of data values contained in x to T. This method is equivalent to convert(AbstractArray{LabeledValue{T, K}, N}}, x).

More Options

The behavior of readstat can be adjusted by passing keyword arguments:

ReadStatTables.readstat — Function

readstat(filepath; kwargs...)

Return a ReadStatTable that collects data (including metadata) from a supported data file located at filepath.

Accepted File Extensions

Stata: .dta.

SAS: .sas7bdat and .xpt.

SPSS: .sav and por.

Keywords

usecols::Union{ColumnSelector, Nothing} = nothing: only collect data from the specified columns (variables); collect all columns if usecols=nothing.

row_limit::Union{Integer, Nothing} = nothing: restrict the total number of rows to be read; read all rows if row_limit=nothing.

row_offset::Union{Integer, Nothing} = nothing: specify the offset for the first row to be read; start from the first row (row_offset=0) if row_offset=nothing.

convert_datetime::Bool = true: convert data from any column with a recognized date/time format to Date or DateTime.

file_encoding::Union{String, Nothing} = nothing: manually specify the file character encoding; need to be an iconv-compatible name.

handler_encoding::Union{String, Nothing} = nothing: manually specify the handler character encoding; default to UTF-8.

The accepted types of values for selecting certain variables (data columns) are shown below:

ReadStatTables.ColumnIndex — Type

ColumnIndex

A type union for values accepted by readstat and ReadStatTable for selecting a column. A column can be selected either with the column name as Symbol or String; or with an integer (Int) index based on the position in a table. See also ColumnSelector.

ReadStatTables.ColumnSelector — Type

ColumnSelector

A type union for values accepted by readstat for selecting a single column or multiple columns. The accepted values can be of type ColumnIndex, a UnitRange of integers, an array or a set of ColumnIndex.

File-level metadata can be obtained without reading the entire data file:

ReadStatTables.readstatmeta — Function

readstatmeta(filepath; kwargs...)

Return a ReadStatMeta that collects file-level metadata without reading the full data from a supported data file located at filepath.

Accepted File Extensions

Stata: .dta.

SAS: .sas7bdat and .xpt.

SPSS: .sav and por.

Keywords

file_encoding::Union{String, Nothing} = nothing: manually specify the file character encoding; need to be an iconv-compatible name.

handler_encoding::Union{String, Nothing} = nothing: manually specify the handler character encoding; default to UTF-8.

1The printed output is generated with PrettyTables.jl.

2The time types Date and DateTime are from the Dates module of Julia.