Table Interface

This page provides further details on the interface of ReadStatTable.

ReadStatTables.ReadStatTable — Type

ReadStatTable <: Tables.AbstractColumns

A Tables.jl-compatible column table that collects data from a Stata, SAS or SPSS file.

Accessing Data in ReadStatTable

Commonly used methods are supported for working with ReadStatTable. As a subtype of Tables.AbstractColumns, ReadStatTable also supports the essential methods defined in Tables.jl.

julia> using ReadStatTables, Tables
julia> tb = readstat("data/sample.dta")5×7 ReadStatTable:
 Row │ mychar    mynum      mydate                dtime         mylabl         ⋯
     │ String  Float64       Date?            DateTime?  Labeled{Int8}  Labele ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │      a      1.1  2018-05-06  2018-05-06T10:10:10           Male         ⋯
   2 │      b      1.2  1880-05-06  1880-05-06T10:10:10         Female         ⋯
   3 │      c  -1000.3  1960-01-01  1960-01-01T00:00:00           Male         ⋯
   4 │      d     -1.4  1583-01-01  1583-01-01T00:00:00         Female         ⋯
   5 │      e   1000.3     missing              missing           Male         ⋯
                                                               2 columns omitted

A column can be accessed either by name or by position via multiple methods:

julia> tb.mynum5-element Vector{Float64}:
     1.1
     1.2
 -1000.3
    -1.4
  1000.3
julia> tb[:mynum]5-element Vector{Float64}:
     1.1
     1.2
 -1000.3
    -1.4
  1000.3
julia> Tables.getcolumn(tb, :mynum)5-element Vector{Float64}:
     1.1
     1.2
 -1000.3
    -1.4
  1000.3
julia> tb[2]5-element Vector{Float64}:
     1.1
     1.2
 -1000.3
    -1.4
  1000.3
julia> Tables.getcolumn(tb, 2)5-element Vector{Float64}:
     1.1
     1.2
 -1000.3
    -1.4
  1000.3

To check whether a column is in a ReadStatTable:

julia> haskey(tb, :mynum)true
julia> haskey(tb, 2)true

To check the number of rows in a ReadStatTable:

julia> Tables.rowcount(tb)5
julia> size(tb, 1)5

To check the number of columns in a ReadStatTable:

julia> length(tb)7
julia> size(tb, 2)7

Iterating a ReadStatTable directly results in iteration across columns:

julia> for col in tb
           println(eltype(col))
       endString
Float64
Union{Missing, Date}
Union{Missing, DateTime}
LabeledValue{Int8}
LabeledValue{Union{Missing, Int8}}
DateTime

Accessing Metadata in ReadStatTable

When calling readstat, a ReadStatMeta object, which collects metadata from the data file, is saved in the ReadStatTable. This object can be retrieved from ReadStatTable via getmeta.

ReadStatTables.ReadStatMeta — Type

ReadStatMeta

A collection of metadata parsed from a data file.

ReadStatTables.getmeta — Function

getmeta(tb::ReadStatTable)

Retrieve the metadata parsed from a data file.

When shown on REPL, a list of the available metadata are printed:

julia> getmeta(tb)ReadStatMeta:
  variable labels:    Dict(:myord => "ordinal", :mynum => "numeric", :mydate => "date", :mychar => "character", :dtime => "datetime", :mytime => "time", :mylabl => "labeled")
  variable formats:   Dict(:myord => "%16.0f", :mynum => "%16.2f", :mydate => "%td", :mychar => "%-1s", :dtime => "%tc", :mytime => "%tcHH:MM:SS", :mylabl => "%16.0f")
  value label names:  Dict(:myord => "myord", :mynum => "", :mydate => "", :mychar => "", :dtime => "", :mytime => "", :mylabl => "mylabl")
  value labels:       Dict{String, Dict{Any, String}}("myord" => Dict(2 => "medium", 3 => "high", 1 => "low"), "mylabl" => Dict(2 => "Female", 1 => "Male"))
  file label:         A test file
  file timestamp:     2021-04-22T21:36:00
  file extension:     .dta

Each field of ReadStatMeta can be accessed either directly from ReadStatMeta or from ReadStatTable via the corresponding accessor function.

ReadStatTables.varlabels — Function

varlabels(m::ReadStatMeta)
varlabels(tb::ReadStatTable)

Retrieve the variable labels as a dictionary indexed by variable names.

ReadStatTables.varformats — Function

varformats(m::ReadStatMeta)
varformats(tb::ReadStatTable)

Retrieve the variable format strings as a dictionary indexed by variable names.

ReadStatTables.val_label_keys — Function

val_label_keys(m::ReadStatMeta)
val_label_keys(tb::ReadStatTable)

Retrieve the names of the collections of value labels applied to each variable as a dictionary indexed by variable names.

ReadStatTables.val_label_dict — Function

val_label_dict(m::ReadStatMeta)
val_label_dict(tb::ReadStatTable)

Retrieve all collections of value labels stored in the data file as a dictionary indexed by the names of the collections. Each collection of value labels itself is again a dictionary that maps the values to the associated labels.

ReadStatTables.filelabel — Function

filelabel(m::ReadStatMeta)
filelabel(tb::ReadStatTable)

Retrieve the label of the data file.

ReadStatTables.filetimestamp — Function

filetimestamp(m::ReadStatMeta)
filetimestamp(tb::ReadStatTable)

Retrieve the time stamp of the data file. (Time zone is not handled.)

ReadStatTables.fileext — Function

fileext(m::ReadStatMeta)
fileext(tb::ReadStatTable)

Retrieve the file extension of the data file.