Date and Time Values

Date and time values in the data files are recognized based on the format of each variable. Most data/time formats can be recognized without user intervention.[1] In case certain date/time formats are not recognized, they can be added easily.

Translating Date and Time Values

For all date/time formats from Stata, SAS and SPSS, the date and time values are stored as the numbers of periods elapsed since a reference date or time point (epoch) chosen by the software. Therefore, knowing the reference data/time and the length of a single period is sufficient for uncovering the represented date/time values for a given format.

The full lists of recognized date/time formats for the statistical software are stored as dictionary keys; while the associated values are tuples of reference date/time and period length.[2] If a variable is in a date/time format that can be found in the dictionary, readstat will handle the conversion to a Julia time type (unless the convert_datetime option prevents it). Otherwise, if a date/time format is not found in the dictionary, no type conversion will be attempted. Additional formats may be added by inserting key-value pairs to the relevant dictionaries.

julia> ReadStatTables.stata_dt_formatsDict{String, Tuple{Union{Date, DateTime}, Dates.Period}} with 7 entries:
  "%tc" => (DateTime("1960-01-01T00:00:00"), Millisecond(1))
  "%tw" => (Date("1960-01-01"), Week(1))
  "%td" => (Date("1960-01-01"), Day(1))
  "%tq" => (Date("1960-01-01"), Month(3))
  "%tm" => (Date("1960-01-01"), Month(1))
  "%th" => (Date("1960-01-01"), Month(6))
  "%ty" => (Date("0000-01-01"), Year(1))
julia> ReadStatTables.sas_dt_formats["MMDDYY"](Date("1960-01-01"), Dates.Day(1))
julia> ReadStatTables.spss_dt_formats["TIME"](DateTime("1582-10-14T00:00:00"), Dates.Second(1))

Translation of the date/time values into a Julia time type is handled by parse_datetime, which is not exported.

ReadStatTables.parse_datetime — Function

parse_datetime(col, epoch::Union{DateTime,Date}, delta::Period, hasmissing::Bool)

Construct a vector of time values of type DateTime or Date by interpreting the elements in col as the number of periods passed since epoch with the length of each period being delta. Returned object is of a type acceptable by ReadStatColumns.

1For Stata, all date/time formats except %tC and %d are supported. The only difference between the %tC format and the %tc format is that %tC takes into account leap seconds while %tc does not. Since the DateTime type in the Dates module does not allow leap seconds, only the %tc format is supported. The %d format that appears in earlier versions of Stata is no longer documented in recent versions.

2For Stata, the reference for date/time value translation is the official Stata documentation. Only the first three characters in the format strings affect the coding. For SAS and SPSS, the reference is pyreadstat/_readstat_parser.pyx.