Date and Time Values

Date and time values in the data files are recognized based on the variable format for each variable. For Stata, all date/time formats except %tC and %d are supported.[1] For SAS and SPSS, only a subset of all possible date/time formats are recognized. Additional date/time formats may be added easily.

Translating Date and Time Values

For all date/time formats from the statistical software, the date and time values are always stored as the numbers of periods passed since a reference date or time point chosen by the software. Therefore, knowing the reference data/time (epoch) and the length of a single period is sufficient for uncovering the represented date/time values for a given format.

The full lists of recognized date/time formats for the statistical software are stored as dictionary keys; while the associated values are tuples of reference date/time and period length.[2] If a variable is in a date/time format that is contained in the dictionary keys, readstat will handle the conversion into a Julia time type (unless the convert_datetime option prevents it). Otherwise, if a date/time format is not found in the dictionary keys, no type conversion will be attempted. Additional formats may be added by inserting key-value pairs to the relevant dictionaries.

julia> ReadStatTables.stata_dt_formats
Dict{String, Tuple{Union{Date, DateTime}, Dates.Period}} with 7 entries:
  "%tc" => (DateTime("1960-01-01T00:00:00"), Millisecond(1))
  "%tw" => (Date("1960-01-01"), Week(1))
  "%td" => (Date("1960-01-01"), Day(1))
  "%tq" => (Date("1960-01-01"), Month(3))
  "%tm" => (Date("1960-01-01"), Month(1))
  "%th" => (Date("1960-01-01"), Month(6))
  "%ty" => (Date("0000-01-01"), Year(1))

julia> ReadStatTables.sas_dt_formats
Dict{String, Tuple{Union{Date, DateTime}, Dates.Period}} with 12 entries:
  "TIME"       => (DateTime("1960-01-01T00:00:00"), Second(1))
  "HHMM"       => (DateTime("1960-01-01T00:00:00"), Second(1))
  "WEEKDATE"   => (Date("1960-01-01"), Day(1))
  "DATETIME20" => (DateTime("1960-01-01T00:00:00"), Second(1))
  "MMDDYY"     => (Date("1960-01-01"), Day(1))
  "YYMMDD10"   => (Date("1960-01-01"), Day(1))
  "TIME20.3"   => (DateTime("1960-01-01T00:00:00"), Second(1))
  "DATETIME"   => (DateTime("1960-01-01T00:00:00"), Second(1))
  "YYMMDD"     => (Date("1960-01-01"), Day(1))
  "DATE9"      => (Date("1960-01-01"), Day(1))
  "DDMMYY"     => (Date("1960-01-01"), Day(1))
  "DATE"       => (Date("1960-01-01"), Day(1))

julia> ReadStatTables.spss_dt_formats
Dict{String, Tuple{Union{Date, DateTime}, Dates.Period}} with 19 entries:
  "TIME"         => (DateTime("1582-10-14T00:00:00"), Second(1))
  "DTIME"        => (DateTime("1582-10-14T00:00:00"), Second(1))
  "TIME5"        => (DateTime("1582-10-14T00:00:00"), Second(1))
  "EDATE"        => (DateTime("1582-10-14T00:00:00"), Second(1))
  "DATE11"       => (DateTime("1582-10-14T00:00:00"), Second(1))
  "DATE8"        => (DateTime("1582-10-14T00:00:00"), Second(1))
  "EDATE8"       => (DateTime("1582-10-14T00:00:00"), Second(1))
  "DATETIME20"   => (DateTime("1582-10-14T00:00:00"), Second(1))
  "JDATE"        => (DateTime("1582-10-14T00:00:00"), Second(1))
  "SDATE"        => (DateTime("1582-10-14T00:00:00"), Second(1))
  "YMDHMS20"     => (DateTime("1582-10-14T00:00:00"), Second(1))
  "TIME11.2"     => (DateTime("1582-10-14T00:00:00"), Second(1))
  "DATETIME"     => (DateTime("1582-10-14T00:00:00"), Second(1))
  "DATETIME23.2" => (DateTime("1582-10-14T00:00:00"), Second(1))
  "EDATE10"      => (DateTime("1582-10-14T00:00:00"), Second(1))
  "DATETIME8"    => (DateTime("1582-10-14T00:00:00"), Second(1))
  "ADATE"        => (DateTime("1582-10-14T00:00:00"), Second(1))
  "DATE"         => (DateTime("1582-10-14T00:00:00"), Second(1))
  "TIME8"        => (DateTime("1582-10-14T00:00:00"), Second(1))

The translation of the date/time values into a Julia time type is handled by parse_datetime (which is not exported).

ReadStatTables.parse_datetime — Function

parse_datetime(col::Vector, epoch::Union{DateTime,Date}, delta::Period)
parse_datetime(col::Vector, epoch::Union{DateTime,Date}, delta::Period, missingvalue)

Construct a vector of time values of type DateTime or Date by interpreting the elements in col as the number of periods passed since epoch with the length of each period being delta. If missingvalue is specified, indices where the elements in col are equal to missingvalue based on isequal are set to be missing no matter what is specified with missingvalue.

1The only difference between the %tC format and the %tc format is that %tC takes into account leap seconds while %tc does not. Since the DateTime type in the Dates module does not allow leap seconds, only the %tc format is supported. The %d format that appears in earlier versions of Stata is no longer documented in recent versions.

2For Stata, the reference for date/time value translation is the official Stata documentation. Only the first three characters in the format strings affect the coding. For SAS and SPSS, the reference is pyreadstat/_readstat_parser.pyx, which only covers a subset of all possible date/time formats documented by official documentation (there are too many of them).