Getting Started

Here is an introduction to the main function of ReadStatTables.jl. For instructions on installation, see Installation.

Loading a Data File

Suppose we have a Stata .dta file located at data/sample.dta. To load this file into Julia, run

julia> using ReadStatTables

julia> tb = readstat("data/sample.dta")
5×7 ReadStatTable:
 Row │ mychar    mynum      mydate                dtime         mylabl         ⋯
     │ String  Float64       Date?            DateTime?  Labeled{Int8}  Labele ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │      a      1.1  2018-05-06  2018-05-06T10:10:10           Male         ⋯
   2 │      b      1.2  1880-05-06  1880-05-06T10:10:10         Female         ⋯
   3 │      c  -1000.3  1960-01-01  1960-01-01T00:00:00           Male         ⋯
   4 │      d     -1.4  1583-01-01  1583-01-01T00:00:00         Female         ⋯
   5 │      e   1000.3     missing              missing           Male         ⋯
                                                               2 columns omitted

Here is how we read the above result:[1]

Variable names from the data file are displayed in the first row.

Element type for each variable is displayed below the corresponding variable name.

The values for each variable are displayed column-wise starting from the third row.

Some additional details to be noted:

If a variable contains any missing value, there is a question mark ? in the displayed element type.

By default, all missing values are treated as missing, a special value in Julia.

The date and time values have been translated into Date and DateTime respectively.[2]

Labels instead of the numeric values are displayed for variables with value labels.

Labeled{Int8} is an abbreviation for LabeledValue{Int8}.

Accessing Individual Objects

A vector of all variable names can be obtained as follows:

julia> columnnames(tb)
7-element Vector{Symbol}:
 :mychar
 :mynum
 :mydate
 :dtime
 :mylabl
 :myord
 :mytime

To retrieve the array containing data for a specific variable:

julia> tb.mylabl
5-element LabeledVector{Int8, LabeledValue{Int8}}:
 1 => Male
 2 => Female
 1 => Male
 2 => Female
 1 => Male

Note

The returned array is exactly the same array holding the data for the table. Therefore, modifying elements in the returned array will also change the data in the table. To avoid such changes, please make a copy of the array first (by calling copy).

Some metadata for the data file are also contained in tb:

julia> getmeta(tb)
ReadStatMeta:
  variable labels:    Dict(:myord => "ordinal", :mynum => "numeric", :mydate => "date", :mychar => "character", :dtime => "datetime", :mytime => "time", :mylabl => "labeled")
  variable formats:   Dict(:myord => "%16.0f", :mynum => "%16.2f", :mydate => "%td", :mychar => "%-1s", :dtime => "%tc", :mytime => "%tcHH:MM:SS", :mylabl => "%16.0f")
  value label names:  Dict(:myord => "myord", :mynum => "", :mydate => "", :mychar => "", :dtime => "", :mytime => "", :mylabl => "mylabl")
  value labels:       Dict{String, Dict{Any, String}}("myord" => Dict(2 => "medium", 3 => "high", 1 => "low"), "mylabl" => Dict(2 => "Female", 1 => "Male"))
  file label:         A test file
  file timestamp:     2021-04-22T21:36:00
  file extension:     .dta

See Table Interface for more complete reference.

Type Conversions

The types provided by ReadStatTables.jl should be sufficient for basic tasks. In case more complicated operations are needed, it is easy to convert the objects into other types.

Converting ReadStatTable

The table returned by readstat is a ReadStatTable. Converting the table into another table type is easy as long as the table type can be constructed with an input following the Tables.jl interface.

For example, to convert the table into a DataFrame from DataFrames.jl, we run

julia> using DataFrames

julia> df = DataFrame(tb)
5×7 DataFrame
 Row │ mychar  mynum    mydate      dtime                mylabl     myord      ⋯
     │ String  Float64  Date?       DateTime?            LabeledV…  LabeledV…  ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │ a           1.1  2018-05-06  2018-05-06T10:10:10  Male       low        ⋯
   2 │ b           1.2  1880-05-06  1880-05-06T10:10:10  Female     medium
   3 │ c       -1000.3  1960-01-01  1960-01-01T00:00:00  Male       high
   4 │ d          -1.4  1583-01-01  1583-01-01T00:00:00  Female     low
   5 │ e        1000.3  missing     missing              Male       missing    ⋯
                                                                1 column omitted

Converting LabeledArray

Variables with value labels are stored in LabeledArrays. To convert a LabeledArray into another array type, we need to determine whether we should keep the values or the labels. If only the labels contain the relevant information, we can make use of the labels function which returns an iterator for the labels. For example, to convert a LabeledArray to a CategoricalArray from CategoricalArrays.jl, we run

julia> using CategoricalArrays

julia> CategoricalArray(labels(tb.mylabl))
5-element CategoricalArray{String,1,UInt32}:
 "Male"
 "Female"
 "Male"
 "Female"
 "Male"

Sometimes, the values have special meanings while the labels are not so important. To obtain an array of the values without the labels, we can call refarray:

julia> refarray(tb.mylabl)
5-element Vector{Int8}:
 1
 2
 1
 2
 1

Alternatively, for a specific element type in the output array, we can call convert:

julia> convert(Vector{Int}, tb.mylabl)
5-element Vector{Int64}:
 1
 2
 1
 2
 1

In the last example, the element type of the output array has become Int.

Note

The array returned by refarray (and by convert if element type is not converted) is exactly the same array underlying the LabeledArray. Therefore, modifying the elements of the array will also modify the values in the original LabeledArray.

More Options

The behavior of readstat can be adjusted by passing keyword arguments.

ReadStatTables.readstat — Function

readstat(filepath::AbstractString; kwargs...)

Return a ReadStatTable that collects data from a supported data file located at filepath.

Accepted File Extensions

Stata: .dta.

SAS: .sas7bdat and .xpt.

SPSS: .sav and por.

Keywords

usecols::Union{ColumnSelector, Nothing} = nothing: only keep data from the specified columns (variables); keep all columns if usecols=nothing.

convert_datetime::Union{Bool, ColumnSelector} = true: convert data from the specified columns to Date or DateTime if they are recorded in supported time formats; if specified as true (false), always (never) convert the data whenever possible.

apply_value_labels::Union{Bool, ColumnSelector} = true: convert data from the specified columns to LabeledArray with their value labels; if specified as true (false), always (never) convert the data whenever possible.

missingvalue = missing: value used to fill any missing value (should be missing unless in special circumstances).

The values used to specify certain variables (columns) must be either a ColumnIndex or a vector of ColumnIndexs.

ReadStatTables.ColumnIndex — Type

ColumnIndex

A type union for values accepted by readstat for selecting a column. A column can be selected either with the column name as Symbol or String; or with an integer index based on the position in a table. See also ColumnSelector.

ReadStatTables.ColumnSelector — Type

ColumnSelector

A type union for values accepted by readstat for selecting either a single column or multiple columns. The accepted values must be either a ColumnIndex or a vector of ColumnIndex.

1The printed output is generated with PrettyTables.jl.

2The time types Date and DateTime are from the Dates module of Julia.