Metadata

File-level metadata associated with a data file are collected in a ReadStatMeta; while variable-level metadata associated with each data column are collected in ReadStatColMetas. These metadata objects are stored in a ReadStatTable along with the data columns and can be accessed via methods compatible with DataAPI.jl.

File-Level Metadata

Each ReadStatTable contains a ReadStatMeta for file-level metadata.

ReadStatTables.ReadStatMeta — Type

ReadStatMeta <: AbstractMetaDict

A collection of file-level metadata associated with a data file processed with ReadStat.

To retrieve the ReadStatMeta from the ReadStatTable:

julia> metadata(tb)ReadStatMeta:
  file label     => A test file
  value labels   => ["mylabl", "myord"]
  timestamp      => 2021-04-22T21:36:00
  file extension => .dta

The value associated with a specific metadata key can be retrieved via:

julia> metadata(tb, "filelabel")"A test file"
julia> metadata(tb, "filelabel", style=true)("A test file", :default)

To obtain a complete list of metadata keys:

julia> metadatakeys(tb)("filelabel", "vallabels", "timestamp", "fileext")

Metadata contained in a ReadStatMeta can be modified, optionally with a metadata style set at the same time:

julia> metadata!(tb, "filelabel", "A file label", style=:note)ReadStatMeta:
  file label     => A file label
  value labels   => ["mylabl", "myord"]
  timestamp      => 2021-04-22T21:36:00
  file extension => .dta

Since ReadStatMeta has a dictionary-like interface, one can also directly work with it:

julia> m = metadata(tb)ReadStatMeta:
  file label     => A file label
  value labels   => ["mylabl", "myord"]
  timestamp      => 2021-04-22T21:36:00
  file extension => .dta
julia> keys(m)KeySet for a ReadStatMeta with 4 entries. Keys:
  "filelabel"
  "vallabels"
  "timestamp"
  "fileext"
julia> m["filelabel"]"A file label"
julia> m["filelabel"] = "A new file label""A new file label"
julia> copy(m)Dict{String, Any} with 4 entries:
  "vallabels" => Dict{String, Dict{Any, String}}("myord"=>Dict(2=>"medium", 3=>…
  "fileext"   => ".dta"
  "filelabel" => "A new file label"
  "timestamp" => DateTime("2021-04-22T21:36:00")

Variable-Level Metadata

A ReadStatColMeta is associated with each data column for variable-level metadata.

ReadStatTables.ReadStatColMeta — Type

ReadStatColMeta <: AbstractMetaDict

A collection of variable-level metadata associated with a data column processed with ReadStat.

To retrieve the ReadStatColMeta for a specified data column contained in a ReadStatTable:

julia> colmetadata(tb, :mylabl)ReadStatColMeta:
  label         => labeled
  format        => %16.0f
  value label   => mylabl
  measure       => 0
  alignment     => 0
  storage width => 1

The value associated with a specific metadata key can be retrieved via:

julia> colmetadata(tb, :mylabl, "label")"labeled"
julia> colmetadata(tb, :mylabl, "label", style=true)("labeled", :default)

To obtain a complete list of metadata keys:

julia> colmetadatakeys(tb, :mylabl)("label", "format", "vallabel", "measure", "alignment", "storagewidth")

Metadata contained in a ReadStatColMeta can be modified, optionally with a metadata style set at the same time:

julia> colmetadata!(tb, :mylabl, "label", "A variable label", style=:note)ColMetaIterator{ReadStatColMeta} with 7 entries:
  :mychar => ReadStatColMeta(character, %-1s)
  :mynum  => ReadStatColMeta(numeric, %16.2f)
  :mydate => ReadStatColMeta(date, %td)
  :dtime  => ReadStatColMeta(datetime, %tc)
  :mylabl => ReadStatColMeta(A variable label, %16.0f)
  :myord  => ReadStatColMeta(ordinal, %16.0f)
  :mytime => ReadStatColMeta(time, %tcHH:MM:SS)

A ReadStatColMeta also has a dictionary-like interface:

julia> m = colmetadata(tb, :mylabl)ReadStatColMeta:
  label         => A variable label
  format        => %16.0f
  value label   => mylabl
  measure       => 0
  alignment     => 0
  storage width => 1
julia> keys(m)KeySet for a ReadStatColMeta with 6 entries. Keys:
  "label"
  "format"
  "vallabel"
  "measure"
  "alignment"
  "storagewidth"
julia> m["label"]"A variable label"
julia> copy(m)Dict{String, Any} with 6 entries:
  "label"        => "A variable label"
  "format"       => "%16.0f"
  "measure"      => 0
  "storagewidth" => 0x0000000000000001
  "alignment"    => 0
  "vallabel"     => "mylabl"

However, it can not be modified directly via setindex!:

julia> m["label"] = "A new label"ERROR: MethodError: no method matching setindex!(::ReadStatColMeta, ::String, ::String)
Closest candidates are:
  setindex!(::AbstractDict, ::Any, ::Any, !Matched::Any, !Matched::Any...) at abstractdict.jl:544

Instead, since the metadata associated with each key are stored consecutively in arrays internally, one may directly access the underlying array for a given metadata key:

ReadStatTables.colmetavalues — Function

colmetavalues(tb::ReadStatTable, key)

Return an array of metadata values associated with key for all columns in tb.

julia> v = colmetavalues(tb, "label")7-element Vector{String}:
 "character"
 "numeric"
 "date"
 "datetime"
 "A variable label"
 "ordinal"
 "time"

Notice that changing any value in the array returned above will affect the corresponding ReadStatColMeta:

julia> colmetadata(tb, :mychar, "label")"character"
julia> v[1] = "char""char"
julia> colmetadata(tb, :mychar, "label")"char"

Metadata Styles

Metadata styles provide additional information on how the metadata should be processed in certain scenarios. ReadStatTables.jl does not require such information. However, specifying metadata styles can be useful when the metadata need to be transferred to some other object (e.g., DataFrame from DataFrames.jl). Packages that implement metadata-related methods compatible with DataAPI.jl are able to recognize the metadata contained in ReadStatTable.

By default, all metadata have the :default style. The user-specified metadata styles are recorded in a Dict based on the keys of metadata:

julia> metastyle(tb)Dict{Symbol, Symbol} with 2 entries:
  :label     => :note
  :filelabel => :note

All metadata associated with keys not listed above are of :default style. To modify the metadata style for those associated with a given key:

julia> metastyle!(tb, "timestamp", :note)Dict{Symbol, Symbol} with 3 entries:
  :timestamp => :note
  :label     => :note
  :filelabel => :note

The same method is also used for variable-specific metadata. However, since the styles are only determined by the metadata keys, metadata associated with the same key always have the same style and hence are not distinguished across different columns.

julia> metastyle!(tb, "label", :note)Dict{Symbol, Symbol} with 3 entries:
  :timestamp => :note
  :label     => :note
  :filelabel => :note
julia> colmetadata(tb, :mychar, "label", style=true)("char", :note)
julia> colmetadata(tb, :mynum, "label", style=true)("numeric", :note)

ReadStatTables.metastyle — Function

metastyle(tb::ReadStatTable, [key::Union{Symbol, AbstractString}])

Return the specified style(s) of all metadata for table tb. If a metadata key is specified, only the style for the associated metadata are returned. By default, the style for all metadata is :default.

The style of metadata is only determined by key and hence is not distinguished across different columns.

ReadStatTables.metastyle! — Function

metastyle!(tb::ReadStatTable, key::Union{Symbol, AbstractString}, style::Symbol)

Set the style of all metadata associated with key to style for table tb.

The style of metadata is only determined by key and hence is not distinguished across different columns.