Value Labels

Value labels from the data files are incorporated into the data columns via a customized array type LabeledArray.

LabeledValue and LabeledArray

LabeledValue and LabeledArray are designed to imitate how variables associated with value labels are represented in the original data files from the statistical software. The element of a LabeledArray is always a LabeledValue.

In general, variables associated with value labels should not be treated as categorical data. Here are some noteworthy distinctions of LabeledArray from an array type designed for categorical data (e.g., CategoricalArray):

Values are never recoded when a LabeledArray is constructed.[1]

It is allowed for some values in a LabeledArray to not have a label.[2]

A label is always a String even when it is associated with missing.

More details are below.

ReadStatTables.LabeledValue — Type

LabeledValue{T}

Value of type T that is associated with a dictionary of value labels. If a value v is not a key in the dictionary, then string(v) is taken as the label. See also LabeledArray.

The value underlying a LabeledValue can be accessed via unwrap. The label can be obtained by converting LabeledValue to String or calling labels (notice the s at the end).

Comparison operators ==, isequal, <, isless and isapprox compare the underlying value of type T. An exception is that when a LabeledValue and a string are compared with ==, the comparison is based on the label.

Examples

julia> lbls = Dict{Union{Int,Missing},String}(0=>"a", 1=>"a");

julia> v0 = LabeledValue(0, lbls)
0 => a

julia> v1 = LabeledValue(1, lbls)
1 => a

julia> vm = LabeledValue(missing, lbls)
missing => missing

julia> v0 == v1
false

julia> v1 == 1
true

julia> v1 == "a"
true

julia> isequal(vm, missing)
true

ReadStatTables.LabeledArray — Type

LabeledArray{V, N, T<:LabeledValue} <: AbstractArray{T, N}

N-dimensional dense array with elements associated with labels.

LabeledArray provides functionality that is similar to what value labels achieve in statistical software such as Stata. When printed to REPL, a LabeledArray just looks like an array of labels. Yet, only the underlying values of type V are stored in an Array. The associated labels are looked up from a dictionary of type Dict{V, String}. If a value v is not a key in the dictionary, then string(v) is taken as the label. The elements of type LabeledValue are only constructed lazily when retrieved.

The array of values underlying a LabeledArray can be accessed with refarray; while an iterator over the labels for each element is returned by labels.

Equality comparison as defined by == involving a LabeledArray only compares the underlying values unless the element type of the other array is a subtype of AbstractString. The labels are used for comparison in the latter case.

Examples

julia> lbls1 = Dict(1=>"a", 2=>"b");

julia> lbls2 = Dict(1.0=>"p", 2.0=>"q");

julia> x = LabeledArray([0, 1, 2], lbls1)
3-element LabeledVector{Int64, LabeledValue{Int64}}:
 0 => 0
 1 => a
 2 => b

julia> y = LabeledArray([0.0, 1.0, 2.0], lbls2)
3-element LabeledVector{Float64, LabeledValue{Float64}}:
 0.0 => 0.0
 1.0 => p
 2.0 => q

julia> x == y
true

julia> x == 0:2
true

julia> x == ["0", "a", "b"]
true

ReadStatTables.LabeledVector — Type

LabeledVector{V, T}

Alias for LabeledArray{V, 1, T}.

Accessing Labels and Values

If only the labels of a LabeledArray are needed, an iterator that maintains the shape of the LabeledArray can be obtained by calling labels. The iterator can be used for either collecting all labels in a different array type or retrieving labels for specific values. On the other hand, if only the values are needed, the labels can be ignored if one directly accesses the underlying array that holds the values.

ReadStatTables.labels — Function

labels(x::LabeledValue)

Return the label associated with x.

labels(x::AbstractArray{<:LabeledValue})

Return an iterator over the labels for each element in x. The returned object is a subtype of AbstractArray with the same size of x.

The iterator can be used to collect the labels while discarding the underlying values.

Examples

julia> x = LabeledArray([1, 2, 3], Dict(1=>"a", 2=>"b"))
3-element LabeledVector{Int64, LabeledValue{Int64}}:
 1 => a
 2 => b
 3 => 3

julia> lbls = labels(x)
3-element ReadStatTables.LabelIterator{LabeledVector{Int64, LabeledValue{Int64}}, 1}:
 "a"
 "b"
 "3"

julia> collect(lbls)
3-element Vector{String}:
 "a"
 "b"
 "3"

julia> CategoricalArray(lbls)
3-element CategoricalArray{String,1,UInt32}:
 "a"
 "b"
 "3"

DataAPI.unwrap — Function

unwrap(x::LabeledValue)

Get the value wrapped by x.

DataAPI.refarray — Function

refarry(x::LabeledArray)
refarray(x::SubArray{<:Any, <:Any, <:LabeledArray})
refarray(x::Base.ReshapedArray{<:Any, <:Any, <:LabeledArray})
refarray(x::AbstractArray{<:LabeledValue})

Return the array of values underlying a LabeledArray.

1The values themselves are sometimes meaningful and should not be treated as reference values.

2In case a label is requested for a value that is not associated with a label, the value is converted to String.